+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added beta suffix to the version in issue reports [#3330].
 - Handling speculative prerender requests.
+- Confirmation modal for inverting allowlist.
 
 ### Fixed